cleanup and bugfixes
